rmove users test, add sync test
